Changed json loading to Search from Results Finalized name changes
