+++ frontend/src/Search.jsx
@@ -1,27 +1,44 @@
-import NameSearch from "./Components/AutoComplete";
-import FeatureSearch from "./Components/FeatureSearch";
+import SearchName from "./Components/SearchName";
+import SearchFeature from "./Components/SearchFeature";
 
-import { useState } from "react";
+import { useState, useEffect } from "react";
 import { Modal } from "./Components/Modal";
 import { useNavigate } from "react-router-dom";
+import axios from "axios";
 
 import "./Styles/App.css";
 import "./Styles/BreakPoints.css";
 
 function Search() {
+	//variable for buttons to change which controlles which search type is shown
 	const [searchType, setSearchType] = useState("text");
+	//String that the warning modal shows | If this is set to any nonempty string the modal pops up
 	const [warning, setWarning] = useState("");
+	//Search Data that is passed to searchResult
 	const [searchResults, setSearchResults] = useState({
+		//Which park should be highlighted first
 		selected: 0,
+		//array of other parks to show
 		searchResults: [],
 	});
+	//from router package that sends data to Search Result
 	let navigate = useNavigate();
-	// console.log("search data", searchResults);
+	const [allParks, setAllParks] = useState([]);
+
+	//Onload gets a list off all the parks from the backend
+	useEffect(() => {
+		axios.get("/parks").then(function (response) {
+			// console.log("pure response", response.data);
+			setAllParks(response.data);
+		});
+	}, []);
+
+	//This block decides what search type to show | Defaults to Name based search
 	let searcher;
 	if (searchType === "features") {
 		searcher = (
 			<div>
-				<FeatureSearch />
+				<SearchFeature />
 			</div>
 		);
 	} else if (searchType === "az") {
@@ -33,7 +50,8 @@ function Search() {
 	} else {
 		searcher = (
 			<div>
-				<NameSearch
+				<SearchName
+					allParks={allParks}
 					setSearchResults={setSearchResults}
 					setWarning={setWarning}
 				/>
@@ -41,9 +59,9 @@ function Search() {
 		);
 	}
 
-	// console.log(searchResults);
-	// console.log(searchResults.length);
+	//Checks to see if user searched, if so sends user to SearchResult with search
 	if (searchResults.searchResults.length) {
+		// console.log(searchResults);
 		navigate("../SearchResult", { state: searchResults });
 	}
